+++ app/period-subdomain/src/main/java/com/intellistart/interviewplanning/appltication/service/PeriodOperations.kt
@@ -5,7 +5,7 @@ import com.google.type.Date
 import com.intellistart.interviewplanning.appltication.validation.PeriodValidator
 import com.intellistart.interviewplanning.domain.exception.PeriodException
 import com.intellistart.interviewplanning.domain.model.Period
-import com.intellistart.interviewplanning.port.PeriodOperationsInPort
+import com.intellistart.interviewplanning.appltication.port.PeriodOperationsInPort
 import org.springframework.stereotype.Service
 import java.time.LocalDate
 import java.time.LocalTime
