+++ app/src/main/java/com/intellistart/interviewplanning/slot/application/port/SlotUpdatesOutPort.kt
@@ -1,0 +1,8 @@
+package com.intellistart.interviewplanning.slot.application.port
+
+import com.intellistart.interviewplanning.output.pubsub.slot.SlotUpdatedEvent
+
+interface SlotUpdatesOutPort {
+
+    fun publishSlotUpdate(slotUpdatedEvent: SlotUpdatedEvent)
+}
