+++ app/src/main/java/com/intellistart/interviewplanning/slot/application/port/SlotsEventOutPort.kt
@@ -1,0 +1,8 @@
+package com.intellistart.interviewplanning.slot.application.port
+
+import com.intellistart.interviewplanning.slot.domain.model.Slot
+
+interface SlotsEventOutPort {
+
+    fun publishEvent(slot: Slot, id: String)
+}
