+++ app/src/main/java/com/intellistart/interviewplanning/slot/infrastructure/adapters/nats/event/SlotUpdatesPublisherNats.kt
@@ -2,14 +2,14 @@ package com.intellistart.interviewplanning.slot.infrastructure.adapters.nats.eve
 
 import com.intellistart.interviewplanning.NatsSubject
 import com.intellistart.interviewplanning.output.pubsub.slot.SlotUpdatedEvent
-import com.intellistart.interviewplanning.slot.port.EventPublisherOutPort
+import com.intellistart.interviewplanning.slot.application.port.SlotUpdatesOutPort
 import io.nats.client.Connection
 import org.springframework.stereotype.Component
 
 @Component
-class EventPublisher(private val connection: Connection) : EventPublisherOutPort {
+class SlotUpdatesPublisherNats(private val connection: Connection) : SlotUpdatesOutPort {
 
-    override fun publishNatsSlotUpdatedEvent(slotUpdatedEvent: SlotUpdatedEvent) {
+    override fun publishSlotUpdate(slotUpdatedEvent: SlotUpdatedEvent) {
         val updateEventSubject = NatsSubject.createSlotEventNatsSubject(slotUpdatedEvent.slotId, "UPDATE")
         connection.publish(updateEventSubject, slotUpdatedEvent.toByteArray())
     }
